requirements.txt
django==2.0
oauthlib==3.1.0
psycopg2==2.8.5
social-auth-core==3.3.3
social-auth-app-django==3.1.0
httplib2 ==0.17.3
djangorestframework ==3.11.0
django-allauth == 0.41.0